move exception into exception package
moving this classes to create an exception factory.
