+++ src/main/java/io/r2dbc/postgresql/PostgresqlResult.java
@@ -17,6 +17,7 @@
 package io.r2dbc.postgresql;
 
 import io.r2dbc.postgresql.codec.Codecs;
+import io.r2dbc.postgresql.exception.PostgresqlServerErrorException;
 import io.r2dbc.postgresql.message.backend.BackendMessage;
 import io.r2dbc.postgresql.message.backend.CommandComplete;
 import io.r2dbc.postgresql.message.backend.DataRow;
