+++ src/test/java/io/r2dbc/postgresql/PostgresqlConnectionFactoryTest.java
@@ -17,9 +17,9 @@
 package io.r2dbc.postgresql;
 
 import com.ongres.scram.client.ScramClient;
-import com.ongres.scram.client.ScramSession;
 import io.r2dbc.postgresql.client.Client;
 import io.r2dbc.postgresql.client.TestClient;
+import io.r2dbc.postgresql.exception.PostgresqlServerErrorException;
 import io.r2dbc.postgresql.message.backend.AuthenticationMD5Password;
 import io.r2dbc.postgresql.message.backend.AuthenticationOk;
 import io.r2dbc.postgresql.message.backend.AuthenticationSASL;
