+++ src/test/java/io/r2dbc/postgresql/PostgresqlResultTest.java
@@ -17,6 +17,7 @@
 package io.r2dbc.postgresql;
 
 import io.r2dbc.postgresql.codec.MockCodecs;
+import io.r2dbc.postgresql.exception.PostgresqlServerErrorException;
 import io.r2dbc.postgresql.message.backend.CommandComplete;
 import io.r2dbc.postgresql.message.backend.DataRow;
 import io.r2dbc.postgresql.message.backend.EmptyQueryResponse;
