+++ src/test/java/io/r2dbc/postgresql/PostgresqlServerErrorExceptionTest.java
@@ -16,6 +16,7 @@
 
 package io.r2dbc.postgresql;
 
+import io.r2dbc.postgresql.exception.PostgresqlServerErrorException;
 import io.r2dbc.postgresql.message.backend.Field;
 import org.junit.jupiter.api.Test;
